Adapt to the latest changes in Unikraft core
This commit updates various headers to match the latest posix-time
interface from Unikraft core, while also introducing new glue code
(`__uk_init_tls.c`) to adapt the library to the new TCB/TLS layout.

Checkpatch-Ignore: EMBEDDED_FUNCTION_NAME
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #28

+++ musl-imported/include/sys/ioctl.h
@@ -140,7 +140,7 @@ struct winsize {
 #define SIOCDEVPRIVATE     0x89F0
 #define SIOCPROTOPRIVATE   0x89E0
 
-int ioctl (int, int, ...);
+int ioctl (int, unsigned long int, ...);
 
 #ifdef __cplusplus
 }
